+++ toolkit/committee-selection/query/src/types/mod.rs
@@ -9,7 +9,7 @@ use scale_info::TypeInfo;
 use serde::{Deserialize, Serialize};
 use sidechain_domain::{CandidateKeys, byte_string::ByteString};
 use sp_core::bytes::to_hex;
-use sp_session_validator_management::CommitteeMember as CommitteeMemberT;
+use sp_session_validator_management::CommitteeMember;
 use std::collections::HashMap;
 
 #[derive(
@@ -31,21 +31,21 @@ pub struct GetCommitteeResponse {
 	/// The sidechain epoch
 	pub sidechain_epoch: u64,
 	/// List of committee members
-	pub committee: Vec<CommitteeMember>,
+	pub committee: Vec<ResponseCommitteeMember>,
 }
 
 impl GetCommitteeResponse {
 	/// Constructor for [GetCommitteeResponse]
-	pub fn new<Member: CommitteeMemberT>(
+	pub fn new<AuthorityId, AuthorityKeys>(
 		sidechain_epoch: sidechain_domain::ScEpochNumber,
-		committee: Vec<Member>,
+		committee: Vec<CommitteeMember<AuthorityId, AuthorityKeys>>,
 	) -> GetCommitteeResponse
 	where
-		Member::AuthorityId: AsRef<[u8]>,
+		AuthorityId: AsRef<[u8]> + Clone,
 	{
 		let committee = committee
 			.into_iter()
-			.map(|member| CommitteeMember::new(member.authority_id()))
+			.map(|member| ResponseCommitteeMember::new(member.authority_id()))
 			.collect();
 		GetCommitteeResponse { sidechain_epoch: sidechain_epoch.0, committee }
 	}
@@ -58,7 +58,10 @@ impl GetCommitteeResponse {
 	where
 		AuthorityId: AsRef<[u8]>,
 	{
-		let committee = committee.into_iter().map(|member| CommitteeMember::new(member)).collect();
+		let committee = committee
+			.into_iter()
+			.map(|member| ResponseCommitteeMember::new(member))
+			.collect();
 		GetCommitteeResponse { sidechain_epoch: sidechain_epoch.0, committee }
 	}
 }
@@ -78,12 +81,12 @@ impl GetCommitteeResponse {
 )]
 #[serde(rename_all = "camelCase")]
 /// Committee member represented by their sidechain pub key
-pub struct CommitteeMember {
+pub struct ResponseCommitteeMember {
 	sidechain_pub_key: String,
 }
 
-impl CommitteeMember {
-	/// Constructor for [CommitteeMember]
+impl ResponseCommitteeMember {
+	/// Constructor for [ResponseCommitteeMember]
 	pub fn new<T: AsRef<[u8]>>(bytes: T) -> Self {
 		Self { sidechain_pub_key: to_hex(bytes.as_ref(), false) }
 	}
@@ -107,12 +110,16 @@ mod tests {
 	use serde_json;
 	use sidechain_domain::ScEpochNumber;
 	use sp_core::ecdsa;
+	use sp_session_validator_management::CommitteeMember;
 
 	#[test]
 	fn get_committee_response_to_json_test() {
 		let input = GetCommitteeResponse::new(
 			ScEpochNumber(4703884),
-			vec![(ecdsa::Public::from([0u8; 33]), 0), (ecdsa::Public::from([255u8; 33]), 1)],
+			vec![
+				CommitteeMember::permissioned(ecdsa::Public::from([0u8; 33]), 0),
+				CommitteeMember::permissioned(ecdsa::Public::from([255u8; 33]), 1),
+			],
 		);
 
 		let json_value = serde_json::to_value(input).expect("Serialization failed");